apps/api/requirements.txt
# Core API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# TinyTroupe Integration
openai>=1.3.8
numpy>=1.24.3
pandas>=2.0.3

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.2
aiofiles>=23.2.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Production
gunicorn>=21.2.0

# TinyTroupe specific dependencies
chevron>=0.14.0
python-docx>=1.1.0

# TinyTroupe local package (editable install)
-e ../../packages/tinytroupe-original